tests/functional/requirements-test.txt
pytest-flask==1.2.0
pytest==7.3.1